refactor: 기능 분리 이동, css명 수정
