+++ src/pages/ComingSoon/ComingSoon.css
@@ -1,4 +1,4 @@
-.coming-soon-container {
+.coming_soon_container {
   width: 375px;
   min-height: 100vh;
   background-color: black;
@@ -8,117 +8,15 @@
   font-family: Arial, sans-serif;
 }
 
-.page-title {
+.page_title {
   color: white;
   font-size: 20px;
   font-weight: bold;
   margin: 16px 12px;
 }
 
-.movie-list {
+.movielist {
   display: flex;
   flex-direction: column;
   gap: 24px;
-}
-
-.movie-card {
-  width: 100%;
-  background-color: black;
-  color: white;
-}
-
-.movie-image {
-  width: 100%;
-  height: auto;
-  display: block;
-}
-
-.movie-info {
-  padding: 12px;
-  background-color: black;
-}
-
-.movie-buttons {
-  display: flex;
-  justify-content: flex-end;
-  gap: 24px;
-  padding-right: 12px;
-  margin-bottom: 8px;
-}
-
-.movie-button-remindme,
-.movie-button-share {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  color: white;
-  cursor: pointer;
-}
-
-.remindme-icon-img,
-.movie-icon-img {
-  width: 24px;
-  height: 24px;
-  object-fit: contain;
-  margin-bottom: 7px;
-}
-
-.movie-button-remindme span,
-.movie-button-share span {
-  margin: 0;
-  padding: 0;
-  line-height: 1;
-  font-size: 11.13px;
-}
-
-.movie-meta {
-  text-align: left;
-  padding: 0 0 8px 0;
-}
-
-.release {
-  font-size: 11.4px;
-  line-height: 9px;
-  color: rgba(255, 255, 255, 0.83);
-  margin-bottom: 4px;
-}
-
-.movie-title {
-  font-size: 18.66px;
-  font-weight: bold;
-  line-height: 14px;
-  color: #ffffff;
-  margin-bottom: 4px;
-}
-
-.movie-overview {
-  font-size: 11.4px;
-  color: rgba(255, 255, 255, 0.83);
-  margin-bottom: 8px;
-  display: -webkit-box;
-  -webkit-line-clamp: 3;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  word-break: keep-all;
-}
-
-.movie-genres {
-  font-size: 11.4px;
-  font-weight: 600;
-  color: #ffffff;
-  line-height: 9px;
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-}
-
-.genre-item {
-  margin-right: 4px;
-}
-
-.dot {
-  color: #ffffffb0;
-  font-weight: bold;
-  margin: 0 6px;
 }
