+++ src/pages/ComingSoon/ComingSoon.jsx
@@ -2,24 +2,25 @@ import React from 'react';
 import './ComingSoon.css';
 import { useMovieData } from '../../hooks/useMovieData';
 import { requests } from '../../services/requests';
-import MovieCard from './MovieCard';
+import MovieCard from '../../components/MovieCard/MovieCard';
 
-export default function ComingSoon() {
+const ComingSoon = () => {
   const { data: movies, loading, error } = useMovieData(requests.fetchUpcoming);
 
-  if (loading) return <div className="coming-soon-container">Loading...</div>;
+  if (loading) return <div className="coming_soon_container">Loading...</div>;
   if (error)
     return (
-      <div className="coming-soon-container error">Error loading data</div>
+      <div className="coming_soon_container error">Error loading data</div>
     );
 
   return (
-    <div className="coming-soon-container">
-      <div className="movie-list">
+    <div className="coming_soon_container">
+      <div className="movie_list">
         {movies.map((movie) => (
           <MovieCard key={movie.id} movie={movie} />
         ))}
       </div>
     </div>
   );
-}
+};
+export default ComingSoon;
